+++ app/src/main/java/com/techfind/myapplication/local/repository/TechfindRepository.kt
@@ -11,8 +11,8 @@ class TechfindRepository {
         name: String,
         email: String,
         password: String,
-        document: Int,
-        cel_number: Int
+        document: Double,
+        cel_number: Double
 
     ) {
         val user = User(
