+++ app/src/main/java/com/techfind/myapplication/server/User.kt
@@ -5,7 +5,7 @@ data class User(
     var name: String? = null,
     var email: String? = null,
     var password: String? = null,
-    var document: Int? = null,
-    var cel_number: Int? = null
+    var document: Double? = null,
+    var cel_number: Double? = null
 )
